pyproject.toml
[project]
name = "gov-daisy"
version = "0.1.0"
description = "Australian government expenditure visualiser data tools"
requires-python = ">=3.8"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.0.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.4",
]
[dependency-groups]
dev = [
    "isort>=5.13.2",
    "ruff>=0.8.4",
    "pytest>=8.3.3",
]

[tool.uv]
# uv-specific config can go here if needed
